requirements.txt
numpy
opencv-python
pypotrace
matplotlib 
pillow 
argparse 
svgwrite
streamlit